Refactor finished
Deacoplation has been done, matching the dependencies with each router, which are grouped by similar functionalities.
