+++ pool.js
@@ -2,15 +2,21 @@ const mysql = require("mysql");
 
 class Pool {
   constructor(host, user, password, database) {
-    this._pool = mysql.createPool({
-      host: host,
-      user: user,
-      password: password,
-      database: database,
-    });
+    if (Pool._pool) {
+      this._pool=Pool._pool;
+    }
+    else {
+      this._pool = mysql.createPool({
+        host: host,
+        user: user,
+        password: password,
+        database: database,
+      });
+    }
+
   }
 
-  get_pool() {
+  static get_pool() {
     return this._pool;
   }
 }
