+++ views/login.ejs
@@ -94,7 +94,7 @@
                 </div>
                 <!-- Link that takes the user to the registration page. -->
                 <div class="my-3">
-                    <a href="/register" id="register_hint">¿Aún no tienes una
+                    <a href="/admin/register" id="register_hint">¿Aún no tienes una
                         cuenta? <br> Regístrate aquí</a>
                 </div>
