+++ views/must_be_login.ejs
@@ -26,7 +26,7 @@
             <div class="row justify-content-center my-3">
                 <div class="col-auto">
                     <button class="border border-3 border-danger rounded-3 btn btn-lg bg-danger bg-opacity-25">
-                        <a href="/login" class="text-danger text-truncate">Ir a la pagina de login</a></button>
+                        <a href="/usuario/login" class="text-danger text-truncate">Ir a la pagina de login</a></button>
                 </div>
             </div>
         </div>
