+++ views/pieces/header.ejs
@@ -55,7 +55,7 @@
                                 ¡Bienvenid@
                                 de nuevo, <%= session.user%>!
                             </p>
-                            <a class="login_text" href="/logout">Logout</a>
+                            <a class="login_text" href="/usuario/logout">Logout</a>
                         </div>
                     </div>
 
@@ -76,7 +76,7 @@
                         </div>
 
                         <div class="d-flex flex-column mx-3 px-1 justify-content-center ">
-                            <a id="a_inbox" href="/correo/inbox">Bandeja de entrada</a>
+                            <a id="a_inbox" href="/usuario/inbox">Bandeja de entrada</a>
                         </div>
                     </div>
                     <% } else { %>
@@ -84,8 +84,8 @@
                         <!-- Login and register buttons. -->
                         <div class="offset-5" id="login_reg_disp_cont">
 
-                            <a class="account_but d-flex " id="log_but" href="/login">Login</a>
-                            <a class="account_but d-flex" href="/register">Registrarse</a>
+                            <a class="account_but d-flex " id="log_but" href="/usuario/login">Login</a>
+                            <a class="account_but d-flex" href="/usuario/register">Registrarse</a>
 
                         </div>
                         <% } %>
@@ -110,16 +110,16 @@
                                         <a href="/index.html" class="nav-link" aria-current="page">Inicio</a>
                                     </li>
                                     <li class="nav-item">
-                                        <a href="/gestion_instalacion" class="nav-link">Gestion Instalación</a>
+                                        <a href="/admin/gestion_instalacion" class="nav-link">Gestion Instalación</a>
                                     </li>
                                     <li class="nav-item">
-                                        <a href="/validar_registro" class="nav-link">Validacion Registro</a>
+                                        <a href="/admin/validar_registro" class="nav-link">Validacion Registro</a>
                                     </li>
                                     <li class="nav-item">
-                                        <a href="/busquedaAvanzada" class="nav-link">Busqueda Avanzada</a>
+                                        <a href="/admin/busquedaAvanzada" class="nav-link">Busqueda Avanzada</a>
                                     </li>
                                     <li class="nav-item">
-                                        <a href="/config_system" class="nav-link">Configuración Sistema</a>
+                                        <a href="/admin/config_system" class="nav-link">Configuración Sistema</a>
                                     </li>
                                 </ul>
                             </div>
