+++ views/registroCompletado.ejs
@@ -28,7 +28,7 @@
                         </div>
 
                         <div class="row mx-md-5 mx-3">
-                            <a href="/login"
+                            <a href="/usuario/login"
                                 class="concrete-dest-res-button">
                                 <button class="register_ok_but">Ir a la ventana <br> de login</button>
                             </a>
